+++ CHANGELOG.md
@@ -7,6 +7,13 @@ writing, this is a rolling-release project without any meaningful versioning
 whatsoever. Tags/releases may be created for the sole purpose of documenting
 major updates to the project.
 
+## 2022-11-12
+
+### Fixed
+
+- Fix iframe transparency
+  ([#785](https://github.com/giscus/giscus/pull/785)).
+
 ## 2022-11-06
 
 ### Fixed